refactor: Split resolvers
